travis: also check whole diff when pull request has several commits
Address a common situation when a pull request is reviewed. Suppose a
PR contains one commit, which causes checkpatch warnings. The author
pushes a new commit to fix the warnings. Travis will still complain
because it checks every commit in the PR, and the author has no way to
make sure the fix is actually correct.
To alleviate this problem, add a call to checkpatch that will process
the whole diff added by the PR branch to the target branch (i.e., git
diff HEAD^1...HEAD^2). In the above example, the Travis status will
still be fail until all commits actually comply with the rules, but the
PR author can check that his/her contribution is OK as a whole.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>