build: include submodule in name of per-library variables
Because a library may be built several times for different submodules
(for example, libutils.a is built differently for kernel and user
mode), the names of the variables that hold library-specific flags
have to uniquely identify the instance of the library that's being
built.

Example use case: to build libutils with flag FOO for the core and
flag BAR for the user mode, add this to lib/libutils/sub.mk:

  ifeq ($(sm),core)
  cppflags-lib-y += -DFOO
  else
  cppflags-lib-y += -DBAR
  endif

Without this fix, the above would fail because all the flags would
be added to all the builds of libutils.

Fixes: 3d34e125a253 ("Add support for $(cflags-lib-y)")
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>