[RFC] Add support for user TA profiling with gprof
This adds the infrastructure to collect profiling information from
the trusted applications running in user mode. Application code needs
to be compiled with the -pg switch, so that the compiler generates
extra code into each function prologue to call the instrumentation
entry point (__gnu_mcount_nc or _mcount depending on the architecture).

libutee collects the raw profiling information, which consists in pairs
of program counters (one is the caller and the other one is the callee)
as well as the number of times this specific arc of the call graph has
been invoked.

A system call is added (syscall_gprof_send) to send the profile data
to tee-supplicant, which will write them to disk. The data follow the
gmon.out format, so they can be processed by the standard gprof tool.

Note that this patch only collects the call graph information; PC
sampling is not implemented. As a result, the gprof output will not
contain any time information. The profile is still useful to understand
the dynamic call flow and obtain exact call counts.

This depends on a tee-supplicant patch.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>