[Update] Add PC sampling capability for TA profiling
Note 1: This commit shall eventually be squashed into the main one
("Add support for user TA profiling with gprof").

Note 2: There is a FIXME in gprof_stop_pc_sampling(): the sampling
frequency is currently hardcoded.

This adds the code to supplement the "gmon.out" data with a PC
histogram buffer. When the TA starts, libutee calls the gprof pseudo-ta
to request PC sampling for the current session. A buffer is allocated
by the TEE core and associated with the session. Each time the CPU is
interrupted in user mode, the core checks if the current session is
being profiled, and if so, the value of the program counter is used to
increment a counter in the profiling buffer. As time passes, this
builds a statistical view of the relative amount of time spent for
each value of the program counter.

The sampling buffer is retrieved by libutee when the session is closed
and sent to tee-supplicant in addition to the call graph information.

With this change, gprof is able to report information on time spent in
each function.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>