[libtomcrypt] Fix clear big number in TEE wrapper
Reset a transient object clear datas of this object.
For big numbers, the corresponding method bn_clear reset the underlying
struct mpa_numbase_struct by filling of zeros its data.
However this struct has metadata, and the call of memset is done on the
begin on the structure, so on the metadata + part of the data, instead
of being done only of the datas.
Fix by zero only datas.

Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Tested-by: Emmanuel MICHEL <emmanuel.michel@st.com> (STM platform)
Signed-off-by: Emmanuel MICHEL <emmanuel.michel@st.com>