[Update] Allocate call graph buffer dynamically, init on demand
- Add the PTA_GPROF_ALLOC operation to the gprof pseudo-ta to allocate
memory from the TEE Core secure DDR pool and map it into the TA address
space. The memory is released when the TA context is destroyed. This
memory is used by libutee to record the function call graph.
- libutee: initialize gprof as needed, i.e., when _mcount_internal() is
invoked for the first time.

This eliminates any runtime overhead when the TA is not instrumented.

TODO: make the memory pageable when CFG_PAGED_USER_TA=y?

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>