build: convert out dir to subdir and clean target nukes subdir
With this we now direct output files to `$O/optee_os_out` instead of `$O` itself.

If you accidentally set, eg, O=~, this stops a `make clean` subsequently
from damaging ~, because it'll delete only the subdir ~/optee_os_out.

There are knock-ons from this for consumers of optee_os output.

1) For hikey build someone bodged the whole optee and arm-trusted-firmware thing together with UEFI build.

Well, I am sure he already feels a distant premonition that was not a good thing in the world.

Part of that involves a direct path reference in uefi-tools to an optee_os build product.

Therefore the patch below on uefi-tools is needed to make it consistent with this patch

```
   diff --git a/opteed-build.sh b/opteed-build.sh
   index 3c759eb..238e316 100755
   --- a/opteed-build.sh
   +++ b/opteed-build.sh
   @@ -103,7 +103,7 @@ function build_platform
                   if [ $VERBOSE -eq 1 ]; then
                           echo "Copying tee.bin to "$EDK2_DIR/Build/$PLATFORM_IMAGE_DIR/$BUILD_PROFILE/FV/""
                   fi
   -               cp -a out/arm-plat-"$TOS_PLATFORM"/core/tee.bin "$EDK2_DIR/Build/$PLATFORM_IMAGE_DIR/$BUILD_PROFILE/FV/"
   +               cp -a out/optee_os_out/arm-plat-"$TOS_PLATFORM"/core/tee.bin "$EDK2_DIR/Build/$PLATFORM_IMAGE_DIR/$BUILD_PROFILE/FV/"
           else
                   return 1
           fi
```

2) As part of your meta-build action several things want to be told TA_DEV_KIT_DIR, so they can get your optee scripts and keys.

After this patch, that goes down one level

 -TA_DEV_KIT_DIR=`pwd`/optee_os/out/arm-plat-hikey/export-ta_arm64
 +TA_DEV_KIT_DIR=`pwd`/optee_os/out/optee_os_out/arm-plat-hikey/export-ta_arm64

3) Your external TA builds are also adapted by this patch so their build products in down one dir

Signed-off-by: Andy Green <andy@warmcat.com>