secure-device-id: map device id to OTP apis
Make the default device ID handlers use the
secure-device-id if it was passed in by DTB.

Whatever length the property has in the DTB
is used for the secure device id, up to
the maximum of 160 bytes.  If the available
secure id is too small, it is repeated up
to the required length.

Signed-off-by: Andy Green <andy@warmcat.com>