core: arm: pager: bugfix linking
Fixes problem when linking and mobj_phys_get_pa() and mobj_mm_get_pa()
ends up in the paged region instead of the unpaged as expected.
Unfortunately the KEEP_PAGER() macro used iwth these functions when
compiling for AArch64 triggers an assert in the linker so it has to be
guarded with #ifndef ARM64.

Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey, pager, GP)
Tested-by: Etienne Carriere <etienne.carriere@linaro.org> (qemu_virt, b2260, pager, GP)
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>