core: generic_entry: call console_init() twice
Prepare for reworking the UART drivers to adopt the usual model where
a device is represented by a structure holding private data (base
address for instance) and a generic part (struct serial_chip).

The console output functions may be used by early init code before the
MMU has been turned on. Therefore, the serial drivers need a way to
access the UART using either a physical or a virtual address. But,
only the platform code knows how to translate the physical address to
a virtual one (it knows the memory type to use in the phys_to_virt()
call). So, this commit gives the platform code the opportunity to
update the driver's private data for the console, by calling
console_init() twice: once before the MMU is turned on, and once
after. A call to console_flush() just before MMU enablement
ensures that no pending data are lost.

The conversion of the UART drivers will be done in later commits.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>