core: pTA SDP memref parameters are reserved to secure clients
Accessing SDP memory from a pseudo TA is a secure service reserved
to the secure clients. pseudo TAs involved in Secure Data Path setups
are expected to provide services for secure clients only: SDP TAs.

SDP memory are default not mapped in the core virtual mapping. When
a pTA is invoked from  a TA, the pTA uses the TA virtual mapping to
access memref parameter buffers. When a pTA is invoked from the non
secure world, the sequence relies on core virtual mapping. such SDP
references would needed to be mapped. This is NOT supported in
current implementation. Mapping SDP memref in the core will be
implemented only once contributors really claim for such a support.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>