plat-hikey: add Secure Data Path support
Define the address and size of the memory area to be used for the SDP
pool when CFG_SECURE_DATA_PATH=y. The bottom 3 MiB of the secure DRAM
are used. The TA RAM at the top is reduced accordingly, and the load
address for OP-TEE is shifted by 3 MiB too. Due to this, this patch
depends on another patch in ARM Trusted Firmware.
The overall reserved memory for OP-TEE is unchanged however.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>