plat-ti: Rename platform context to platform boot arguments
Currently the non-secure context is passed in from our initial secure
software as part of the OP-TEE load process. This passed-in data will
not only contain the non-secure context but also any additional data we
may need to give to OP-TEE. Rename these structures and group the context
data into a struct for future expansion.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>