libutee: Remove buffering for AES GCM
GCM mode of AES algorithm is essentially a stream cipher. It means
that there is no requirements for input data size related to block
size. And every time when AEUpdate is called input data must be
processed and output immediately generated. That is why
it is incorrect to apply for AES GCM the input buffering on
AEUpdate call when size is not multiple of block size.

Signed-off-by: Bogdan Liulko <bogdan.liulko@globallogic.com>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey)
Tested-by: Bogdan Liulko <bogdan.liulko@globallogic.com> (R-Car)
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>