Fix of the _core_is_buffer_intersect check
Function _core_is_buffer_intersect() was returning false in case is_valid_conf_and_notnull_size() returns false. Which means that if buffer is zero length or if there is an overflow due to very big length of the buffer the function _core_is_buffer_intersect() will return false indicating that there is no intersection which is not correct since in case of an overflow the buffers still can intersect.