libutee: Remove restrictions for unpadded sign and verify RSA modes
According to Global Platforms Specifications for algorithm
TEE_ALG_RSA_NOPAD support of modes TEE_MODE_SIGN and
TEE_MODE_VEIFY is not required. But it is not forbidden to
implement wider range of modes if it is possible. Cryptographic
library that is used supports sign and verify operations for
RSA algorithm without padding. That is why restrictions for these
modes can be removed.

Change-Id: I91ebb56eb55ff692cdc743c8e4857ff70f024b6a
Signed-off-by: Bogdan Liulko <bogdan.liulko@globallogic.com>