arm.mk: export-ta_arm32: Set default value for CROSS_COMPILE
When building 32-bit optee_os, CROSS_COMPILE32 defaults to CROSS_COMPILE
if not set, and CROSS_COMPILE defaults to arm-linux-gnueabihf- if not
set.

When building 32-bit optee_test, CROSS_COMPILE32 also defaults to
CROSS_COMPILE if not set, but if CROSS_COMPILE is not set, there is no
default value, causing the native compiler to be used and generating
errors and confusion.

This patch makes the behavior and expectations consistent across both
optee_os and optee_test.

Signed-off-by: Victor Chong <victor.chong@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>