Import zlib v1.2.11
Import the decompression code from zlib v1.2.11. From the project's
README:

"zlib 1.2.11 is a general purpose data compression library. [...] The
data format used by the zlib library is described by RFCs (Request for
Comments) 1950 to 1952 in the files rfc1950 (zlib format), rfc1951
(deflate format) and rfc1952 (gzip format)."

This code will be used in a later commit to decompress early TAs.
Only the inflate() function is needed, and the library is configured
without gzip support. The source files that are not required
for inflate() are left aside.

The library is licensed under a permissive license, see `zlib.h`.

Link: http://tools.ietf.org/html/rfc1950
Link: http://tools.ietf.org/html/rfc1951
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>