Travis: add -p to mkdir $HOME/bin
Travis are currently migrating Linux distributions from Ubuntu Precise
(12.04) to Trusty (14.04). It seems that $HOME/bin already exists in
the new images, which causes an error as we do mkdir $HOME/bin.
Add a -p so that the directory creation doesn't fail when bin
exists. This allows the script to run on both Precise and Trusty.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Igor Opaniuk <igor.opaniuk@linaro.org>